requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.115.12
fonttools==4.58.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-pubsub==2.29.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
idna==3.10
importlib_metadata==8.6.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
zipp==3.21.0
google-cloud-storage==2.14.0
python-dotenv>=0.19.0
minio>=7.2.0
pandas==2.2.2 
